updates from self-code review
